(feat) Working on saving spawners and its data in the  map file

+++ include/States/EditorModes/EnemyEditorMode.h
@@ -22,7 +22,7 @@ private:
     short type;
     short amount;
     short timeToSpawn;
-    float maxDistance;
+    int maxDistance;
 
     /* INITIALIZERS ============================================================================== */
